Fixed sonar

+++ src/main/java/g3401_3500/s3461_check_if_digits_are_equal_in_string_after_operations_i/Solution.java
@@ -9,8 +9,8 @@ public class Solution {
         int k = ch.length - 1;
         while (k != 1) {
             for (int i = 0; i < k; i++) {
-                int a = (int) ch[i] - 48;
-                int b = (int) ch[i + 1] - 48;
+                int a = ch[i] - 48;
+                int b = ch[i + 1] - 48;
                 int d = (a + b) % 10;
                 char c = (char) (d + '0');
                 ch[i] = c;
