+++ src/main/java/g3401_3500/s3464_maximize_the_distance_between_points_on_a_square/Solution.java
@@ -15,7 +15,7 @@ public class Solution {
             if (y == 0) {
                 c = x;
             } else if (x == side) {
-                c = side + y;
+                c = side + (long) y;
             } else if (y == side) {
                 c = 2L * side + (side - x);
             } else {
